Add BaseBubbleBuilderHook
